+++ pkg/lvsdk/id.go
@@ -7,7 +7,7 @@ import (
 )
 
 type idDso struct {
-	count  uint64
+	count  uint
 	prefix string
 	mutex  *sync.Mutex
 }
@@ -31,7 +31,7 @@ func (id *idDso) Next() string {
 	return fmt.Sprintf("%s-%d-%d", id.prefix, count, now)
 }
 
-func (id *idDso) next() uint64 {
+func (id *idDso) next() uint {
 	defer id.mutex.Unlock()
 	id.mutex.Lock()
 	id.count++
